Added support for performance profile status
This commit includes the implementation of this PR on the hypershift side:
 https://github.com/openshift/enhancements/pull/1619

Up until now, there was no defined resolution over how PerformanceProfile
status conditions with the status of the different components created
from the PerformanceProfile (MachineConfig, KubeletConfig, Tuned) are handled and exposed.
To resolve this, a custom ConfigMap is added and will be used as a middleware object.
It will hold the updated status populated by the performance profile controller
and will be watched by the NodePool controller, which will calculate an overview
of the current status provided and reflect it under NodePool.status.condition.

The resolution described above consist of the following changes:
* Added a new condition "PerformanceProfileAppliedSuccessfully" to the NodePool conditions API.
* Added a watch to the specific performance profile status ConfigMap object that will trigger the reconcile loop.
* Added the HandlePerformanceProfileStatus function to calculate and set the PerformanceProfileAppliedSuccessfully based on the possible performance profile statuses (available, progressing and degraded).

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>